fix virtualbox host resolver
We need to force VirtualBox to use the host DNS API to resolve names instead of passing DNS requests to an external DNS server. In this way, the local names stored in the local /etc/hosts file (C:\Windows\System32\drivers\etc\hosts in windows) of the VirtualBox host will be used.